Broke HttpHandlerSelector into multiple classes.
