+++ src/test/java/HandleGetRequestTest.java
@@ -2,9 +2,9 @@ import com.td.HttpServer.*;
 import java.util.*;
 import java.io.IOException;
 
-public class HttpHandlerSelectorTest extends junit.framework.TestCase {
+public class HandleGetRequestTest extends junit.framework.TestCase {
 
-  HttpHandlerSelector handler;
+  HandleGetRequest getRequest;
   HttpRequest request;
   HttpResponse response;
 
@@ -18,48 +18,48 @@ public class HttpHandlerSelectorTest extends junit.framework.TestCase {
 
   protected void setUp() {
     MockFileIO mockFileIO = new MockFileIO();
-    handler = new HttpHandlerSelector(mockFileIO);
+    getRequest = new HandleGetRequest(mockFileIO);
   }
 
   public void testGenerateOkResponseDefaultPath() {
     request = new HttpRequest("GET / HTTP/1.1");
-    response = handler.generateResponse(request);
+    response = getRequest.generateResponse(request);
     assertEquals("HTTP/1.1 200 OK", response.responseLine());
   }
 
   public void testGenerateOkResponseValidFile() {
     request = new HttpRequest("GET /something.txt HTTP/1.1");
-    response = handler.generateResponse(request);
+    response = getRequest.generateResponse(request);
     assertEquals("HTTP/1.1 200 OK", response.responseLine());
   }
 
   public void testGenerateNotFoundResponseforInvalidFile() {
     request = new HttpRequest("GET /throwIOException HTTP/1.1");
-    response = handler.generateResponse(request);
+    response = getRequest.generateResponse(request);
     assertEquals("HTTP/1.1 404 Not Found", response.responseLine());
   }
 
   public void testContentTypeIsTextHtmlForTxtFile() {
     request = new HttpRequest("GET /something.txt HTTP/1.1");
-    response = handler.generateResponse(request);
+    response = getRequest.generateResponse(request);
     assertTrue(response.headers().contains("text/html"));
   }
 
   public void testContentTypeIsImageJpegForJpg() {
     request = new HttpRequest("GET /something.jpg HTTP/1.1");
-    response = handler.generateResponse(request);
+    response = getRequest.generateResponse(request);
     assertTrue(response.headers().contains("image/jpeg"));
   }
 
   public void testContentTypeIsImageGifForGif() {
     request = new HttpRequest("GET /something.gif HTTP/1.1");
-    response = handler.generateResponse(request);
+    response = getRequest.generateResponse(request);
     assertTrue(response.headers().contains("image/gif"));
   }
 
   public void testContentTypeIsPdfForPdf() {
     request = new HttpRequest("GET /something.pdf HTTP/1.1");
-    response = handler.generateResponse(request);
+    response = getRequest.generateResponse(request);
     assertTrue(response.headers().contains("application/pdf"));
   }
 }
